remove internal references from external API
