+++ nexus/types/src/external_api/params.rs
@@ -254,7 +254,7 @@ pub struct RouteSelector {
 
 // Silos
 
-/// Create-time parameters for a [`Silo`](crate::external_api::views::Silo)
+/// Create-time parameters for a `Silo`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct SiloCreate {
     #[serde(flatten)]
@@ -270,11 +270,11 @@ pub struct SiloCreate {
     ///
     /// Note that if configuring a SAML based identity provider,
     /// group_attribute_name must be set for users to be considered part of a
-    /// group. See [`SamlIdentityProviderCreate`] for more information.
+    /// group. See `SamlIdentityProviderCreate` for more information.
     pub admin_group_name: Option<String>,
 }
 
-/// Create-time parameters for a [`User`](crate::external_api::views::User)
+/// Create-time parameters for a `User`
 #[derive(Clone, Deserialize, Serialize, JsonSchema)]
 pub struct UserCreate {
     /// username used to log in
@@ -645,14 +645,14 @@ where
 
 // PROJECTS
 
-/// Create-time parameters for a [`Project`](crate::external_api::views::Project)
+/// Create-time parameters for a `Project`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct ProjectCreate {
     #[serde(flatten)]
     pub identity: IdentityMetadataCreateParams,
 }
 
-/// Updateable properties of a [`Project`](crate::external_api::views::Project)
+/// Updateable properties of a `Project`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct ProjectUpdate {
     #[serde(flatten)]
@@ -661,8 +661,7 @@ pub struct ProjectUpdate {
 
 // NETWORK INTERFACES
 
-/// Create-time parameters for an
-/// [`InstanceNetworkInterface`](omicron_common::api::external::InstanceNetworkInterface).
+/// Create-time parameters for an `InstanceNetworkInterface`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct InstanceNetworkInterfaceCreate {
     #[serde(flatten)]
@@ -675,8 +674,7 @@ pub struct InstanceNetworkInterfaceCreate {
     pub ip: Option<IpAddr>,
 }
 
-/// Parameters for updating an
-/// [`InstanceNetworkInterface`](omicron_common::api::external::InstanceNetworkInterface).
+/// Parameters for updating an `InstanceNetworkInterface`
 ///
 /// Note that modifying IP addresses for an interface is not yet supported, a
 /// new interface must be created instead.
@@ -704,8 +702,7 @@ pub struct InstanceNetworkInterfaceUpdate {
 
 // CERTIFICATES
 
-/// Create-time parameters for a
-/// [`Certificate`](crate::external_api::views::Certificate)
+/// Create-time parameters for a `Certificate`
 #[derive(Clone, Deserialize, Serialize, JsonSchema)]
 pub struct CertificateCreate {
     /// common identifying metadata
@@ -731,9 +728,7 @@ impl std::fmt::Debug for CertificateCreate {
 
 // IP POOLS
 
-/// Create-time parameters for an IP Pool.
-///
-/// See [`IpPool`](crate::external_api::views::IpPool)
+/// Create-time parameters for an `IpPool`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct IpPoolCreate {
     #[serde(flatten)]
@@ -819,7 +814,7 @@ pub enum ExternalIpCreate {
     // TODO: Add floating IPs: https://github.com/oxidecomputer/omicron/issues/1334
 }
 
-/// Create-time parameters for an [`Instance`](omicron_common::api::external::Instance)
+/// Create-time parameters for an `Instance`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct InstanceCreate {
     #[serde(flatten)]
@@ -933,7 +928,7 @@ impl JsonSchema for UserData {
     }
 }
 
-/// Migration parameters for an [`Instance`](omicron_common::api::external::Instance)
+/// Migration parameters for an `Instance`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct InstanceMigrate {
     pub dst_sled_id: Uuid,
@@ -973,23 +968,23 @@ pub struct InstanceSerialConsoleData {
 
 // VPCS
 
-/// Create-time parameters for a [`Vpc`](crate::external_api::views::Vpc)
+/// Create-time parameters for a `Vpc`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct VpcCreate {
     #[serde(flatten)]
     pub identity: IdentityMetadataCreateParams,
 
-    /// The IPv6 prefix for this VPC.
+    /// The IPv6 prefix for this VPC
     ///
     /// All IPv6 subnets created from this VPC must be taken from this range,
-    /// which sould be a Unique Local Address in the range `fd00::/48`. The
+    /// which should be a Unique Local Address in the range `fd00::/48`. The
     /// default VPC Subnet will have the first `/64` range from this prefix.
     pub ipv6_prefix: Option<Ipv6Net>,
 
     pub dns_name: Name,
 }
 
-/// Updateable properties of a [`Vpc`](crate::external_api::views::Vpc)
+/// Updateable properties of a `Vpc`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct VpcUpdate {
     #[serde(flatten)]
@@ -997,7 +992,7 @@ pub struct VpcUpdate {
     pub dns_name: Option<Name>,
 }
 
-/// Create-time parameters for a [`VpcSubnet`](crate::external_api::views::VpcSubnet)
+/// Create-time parameters for a `VpcSubnet`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct VpcSubnetCreate {
     #[serde(flatten)]
@@ -1018,7 +1013,7 @@ pub struct VpcSubnetCreate {
     pub ipv6_block: Option<Ipv6Net>,
 }
 
-/// Updateable properties of a [`VpcSubnet`](crate::external_api::views::VpcSubnet)
+/// Updateable properties of a `VpcSubnet`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct VpcSubnetUpdate {
     #[serde(flatten)]
@@ -1027,14 +1022,14 @@ pub struct VpcSubnetUpdate {
 
 // VPC ROUTERS
 
-/// Create-time parameters for a [`VpcRouter`](crate::external_api::views::VpcRouter)
+/// Create-time parameters for a `VpcRouter`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct VpcRouterCreate {
     #[serde(flatten)]
     pub identity: IdentityMetadataCreateParams,
 }
 
-/// Updateable properties of a [`VpcRouter`](crate::external_api::views::VpcRouter)
+/// Updateable properties of a `VpcRouter`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct VpcRouterUpdate {
     #[serde(flatten)]
@@ -1043,7 +1038,7 @@ pub struct VpcRouterUpdate {
 
 // VPC ROUTER ROUTES
 
-/// Create-time parameters for a [`omicron_common::api::external::RouterRoute`]
+/// Create-time parameters for a `RouterRoute`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct RouterRouteCreate {
     #[serde(flatten)]
@@ -1052,7 +1047,7 @@ pub struct RouterRouteCreate {
     pub destination: RouteDestination,
 }
 
-/// Updateable properties of a [`omicron_common::api::external::RouterRoute`]
+/// Updateable properties of a `RouterRoute`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct RouterRouteUpdate {
     #[serde(flatten)]
@@ -1137,7 +1132,7 @@ pub enum DiskSource {
     ImportingBlocks { block_size: BlockSize },
 }
 
-/// Create-time parameters for a [`Disk`](omicron_common::api::external::Disk)
+/// Create-time parameters for a `Disk`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct DiskCreate {
     /// common identifying metadata
@@ -1210,8 +1205,7 @@ pub struct Distribution {
     pub version: String,
 }
 
-/// Create-time parameters for an
-/// [`GlobalImage`](crate::external_api::views::GlobalImage)
+/// Create-time parameters for a `GlobalImage`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct GlobalImageCreate {
     /// common identifying metadata
@@ -1228,8 +1222,7 @@ pub struct GlobalImageCreate {
     pub source: ImageSource,
 }
 
-/// Create-time parameters for an
-/// [`Image`](crate::external_api::views::Image)
+/// Create-time parameters for an `Image`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct ImageCreate {
     /// common identifying metadata
@@ -1251,7 +1244,7 @@ pub struct ImageCreate {
 
 // SNAPSHOTS
 
-/// Create-time parameters for a [`Snapshot`](crate::external_api::views::Snapshot)
+/// Create-time parameters for a `Snapshot`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct SnapshotCreate {
     /// common identifying metadata
@@ -1275,7 +1268,7 @@ pub struct OptionalGroupSelector {
 // for creating them internally as we use for types that can be created in the
 // external API.
 
-/// Create-time parameters for a [`UserBuiltin`](crate::external_api::views::UserBuiltin)
+/// Create-time parameters for a `UserBuiltin`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct UserBuiltinCreate {
     #[serde(flatten)]
@@ -1293,7 +1286,7 @@ pub struct UserBuiltinSelector {
 // and so have an implicit silo user ID which must be passed seperately
 // to the creation routine. Note that this disagrees with RFD 44.
 
-/// Create-time parameters for an [`SshKey`](crate::external_api::views::SshKey)
+/// Create-time parameters for an `SshKey`
 #[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
 pub struct SshKeyCreate {
     #[serde(flatten)]
