+++ nexus/types/src/external_api/shared.rs
@@ -28,8 +28,7 @@ pub use omicron_common::address::{IpRange, Ipv4Range, Ipv6Range};
 // is Nexus.  We should have some kinds of config that lives in the database.
 pub const MAX_ROLE_ASSIGNMENTS_PER_RESOURCE: usize = 64;
 
-/// Client view of a [`Policy`], which describes how this resource may be
-/// accessed
+/// Policy for a particular resource
 ///
 /// Note that the Policy only describes access granted explicitly for this
 /// resource.  The policies of parent resources can also cause a user to have
@@ -66,8 +65,8 @@ where
 /// Describes the assignment of a particular role on a particular resource to a
 /// particular identity (user, group, etc.)
 ///
-/// The resource is not part of this structure.  Rather, [`RoleAssignment`]s are
-/// put into a [`Policy`] and that Policy is applied to a particular resource.
+/// The resource is not part of this structure.  Rather, `RoleAssignment`s are
+/// put into a `Policy` and that Policy is applied to a particular resource.
 #[derive(Clone, Debug, Deserialize, Eq, PartialEq, Serialize, JsonSchema)]
 #[schemars(rename = "{AllowedRoles}RoleAssignment")]
 pub struct RoleAssignment<AllowedRoles> {
